Basic turn management through server
